requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
flask-mail>=0.9.1
flask-limiter>=3.5.0
flask-session>=0.5.0
werkzeug>=3.0.1
requests>=2.31.0
plotly>=5.18.0
pandas>=2.1.0
gunicorn>=21.2.0
pytest>=7.4.0
python-jose>=3.3.0
bcrypt>=4.0.1
pytest-cov>=4.1.0
black>=23.9.1
filelock>=3.12.4
cryptography>=41.0.4
python-magic>=0.4.24
email-validator>=2.0.0
Flask-WTF>=1.1.1
Flask-Talisman>=1.0.0
